refactor: wrap quiz admin in server page
